Move Jitted block to separate module
